% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contamination_level.R
\name{calculate_contamination_level}
\alias{calculate_contamination_level}
\title{Calculate the sample contamination status and contamination level}
\usage{
calculate_contamination_level(variant.level.matrix)
}
\arguments{
\item{variant.level.matrix}{The function inputs a matrix of variant-level contamination 
information generated by \code{\link{classification}};}
}
\value{
The function outputs a matrix of sample-level contamination information.
}
\description{
This function Calculate contamination status and contamination level
by contamination-derived variants.
}
\examples{
variant.level.matrix <- data.frame("Sample" = c("SampleA","SampleB"),
                                   "Variant_allele_frequency" = c(0.9850, 0.1036),
                                   "Predict.contamination.status" = c(0, 0),
                                   "Probability.of.contamination.origin" = c(0.1025, 0.3655))
calculate_contamination_level(variant.level.matrix = variant.level.matrix )
}
