% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Detect sample cross-contamination, evaluate contamination level, and identify
contamination-derived variants}
\usage{
classification(features)
}
\arguments{
\item{features}{The features extracted by \code{\link{extract_features}};}
}
\value{
The function outputs a list: the first element is a matrix of sample-level
contamination results. The second list element is a matrix of variant-level contamination results.
}
\description{
This function identifies the contamination-derived mtDNA variants by
random forest and detects sample contamination status and sample contamination level.
}
\examples{
features <- data.frame("Sample"=c("SampleA","SampleB"),
                       "Variant_allele_frequency"=c(0.9835,0.1036),
                       "Homoplasmic_status" = c(0,0),
                       "dbSNP_status" = c(1, 0),
                       "Number_of_contamination_derived_variant" = c(4, 0),
                       "Variant_frequency_in_MITOMAP" = c(0.439734004,0.213280188),
                       "Haplogroup_defining_variants" = c(1, 0),
                       "Heteroplasmic_variant_number" = c(6,3),
                       "CV_of_heteroplasmic_variant_allele_frequency" = c(0.373981559, 0.330185292),
                       "Sample_contamination_status" = c(1, 0),
                       "Predicted_contamination_status_for_major_heteroplasmic_variants" = c(0, 0),
                       "Predicted_contamination_status_for_minor_heteroplasmic_variants" = c(0, 1))
result <- classification(features = features)
}
