% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoDataclean.R
\name{mitoDataclean}
\alias{mitoDataclean}
\title{mitoDataclean pipeline.}
\usage{
mitoDataclean(variants, visualization = FALSE)
}
\arguments{
\item{variants}{The variant information used as input for feature extraction;
this argument is required}

\item{visualization}{Should the plot of probability of contamination origin
for all variants in a sample be output?
this argument is optional}
}
\value{
The sample-level and variant-level contamination information.
}
\description{
A machine learning approach to retrieve genuine mtDNA mutations from sample
contamination-derived chaos
}
\details{
Calls the functions to run each of the three steps of the pipeline
(extract features, classification, visualization) with the specified parameters.
See the individual function documentation for more details and required arguments.
Required steps: \code{\link{extract_features}}, \code{\link{classification}}.
\code{\link{visualization}}.
}
\examples{
data(variants)
result<-mitoDataclean(variants=variants)
variant.level<-result$variant.level
sample.level<-result$sample.level
}
