% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Plot the classification results against variant allele frequency and probability of contamination origin}
\usage{
visualization(variant.level.matrix, verbose)
}
\arguments{
\item{variant.level.matrix}{matrix containing variant allele frequency and probability of contamination origin, 
as calculated by \code{\link{classification}}}

\item{verbose}{Should the plot be output? default is FALSE.}
}
\value{
plots for each sample(returned silently)
}
\description{
Creates the scatter plots by variant allele frequency and 
probability of contamination origin for each sample.
}
\examples{
variant.level.matrix <- data.frame("Sample" = c("SampleA","SampleB"),
                                   "Variant_allele_frequency" = c(0.9850, 0.1036),
                                   "Predict.contamination.status" = c(0, 0),
                                   "Probability.of.contamination.origin" = c(0.1025, 0.3655))
visualization(variant.level.matrix, verbose=FALSE)
}
